embassy_time_queue_driver
Trait TimerQueue

pub trait TimerQueue {
    // Required method
    fn schedule_wake(&'static self, at: u64, waker: &Waker);
}

Timer queue

Required Methods

fn schedule_wake(&'static self, at: u64, waker: &Waker)

Schedules a waker in the queue to be awoken at moment at. If this moment is in the past, the waker might be awoken immediately.